+++ tests/test_serialisation.py
@@ -1,9 +1,10 @@
 import unittest
-from refract.json import JSONSerialiser
-from refract import Element, String, Array, Metadata, KeyValuePair
+from refract.json import JSONSerialiser, CompactJSONSerialiser
+from refract import (Element, String, Number, Boolean, Null, Array, Metadata,
+                     KeyValuePair)
 
 
-class SerialisationTests(unittest.TestCase):
+class JSONSerialisationTests(unittest.TestCase):
     def setUp(self):
         self.serialiser = JSONSerialiser()
 
@@ -198,3 +199,62 @@ class SerialisationTests(unittest.TestCase):
                 }
             }
         })
+
+
+class CompactJSONSerialisationTests(unittest.TestCase):
+    def setUp(self):
+        self.serialiser = CompactJSONSerialiser()
+
+    def test_serialise_string_element(self):
+        payload = self.serialiser.serialise(String(content='Hello'))
+        self.assertEqual(payload, '["string", null, null, "Hello"]')
+
+    def test_serialise_number_element(self):
+        payload = self.serialiser.serialise(Number(content=2))
+        self.assertEqual(payload, '["number", null, null, 2]')
+
+    def test_serialise_boolean_element(self):
+        payload = self.serialiser.serialise(Boolean(content=True))
+        self.assertEqual(payload, '["boolean", null, null, true]')
+
+    def test_serialise_null_element(self):
+        payload = self.serialiser.serialise(Null())
+        self.assertEqual(payload, '["null", null, null, null]')
+
+    def test_serialise_element_element(self):
+        payload = self.serialiser.serialise(Element('test',
+                                            content=String(content='value')))
+        self.assertEqual(payload, '["test", null, null, '
+                                  '["string", null, null, "value"]]')
+
+    def test_serialise_array_element(self):
+        payload = self.serialiser.serialise(Element('array',
+                                            content=[String(content='value')]))
+        self.assertEqual(payload, '["array", null, null, '
+                                  '[["string", null, null, "value"]]]')
+
+    def test_serialise_key_value_pair_element(self):
+        pair = KeyValuePair(
+            key=String(content='name'),
+            value=String(content='Doe')
+        )
+        payload = self.serialiser.serialise(Element('member', content=pair))
+
+        self.assertEqual(payload, '["member", null, null, ["pair", '
+                                  '["string", null, null, "name"], '
+                                  '["string", null, null, "Doe"]'
+                                  ']]')
+
+    def test_serialise_meta(self):
+        element = String(content='Hello')
+        element.title = 'Title'
+        payload = self.serialiser.serialise(element)
+        self.assertEqual(payload, '["string", {"title": ["string", null, '
+                                  'null, "Title"]}, null, "Hello"]')
+
+    def test_serialise_attributes(self):
+        element = String(content='Hello')
+        element.attributes = {'contentType': String(content='text')}
+        payload = self.serialiser.serialise(element)
+        self.assertEqual(payload, '["string", null, {"contentType": '
+                                  '["string", null, null, "text"]}, "Hello"]')